adjust int32_t values in slashcommand to int64_t

+++ include/dpp/commandhandler.h
@@ -52,7 +52,7 @@ struct CoreExport resolved_user {
  * @brief Represents a received parameter.
  * We use variant so that multiple non-related types can be contained within.
  */
-typedef std::variant<std::monostate, std::string, dpp::role, dpp::channel, dpp::resolved_user, int32_t, bool> command_parameter;
+typedef std::variant<std::monostate, std::string, dpp::role, dpp::channel, dpp::resolved_user, int64_t, bool> command_parameter;
 
 /**
  * @brief Parameter types when registering a command.
